fix --turbo
