+++ ark/fuma/app/docs-og/[...slug]/route.tsx
@@ -1,4 +1,3 @@
-import { fromPackageRoot } from "@ark/fs"
 import { generateOGImage } from "fumadocs-ui/og"
 import { metadataImage } from "../../../lib/metadata"
 
@@ -6,23 +5,23 @@ export const GET = metadataImage.createAPI(page =>
 	generateOGImage({
 		title: page.data.title === "ArkType" ? "ArkType" : "ArkType Docs",
 		description: page.data.title === "ArkType" ? "" : page.data.title,
-		site: "ArkType",
-		font: {
-			title: {
-				families: ["Raleway"],
-				weight: "Bold",
-				size: 100
-			},
-			description: {
-				families: ["Raleway"],
-				weight: "SemiBold",
-				size: 40
-			}
-		},
-		fonts: [
-			fromPackageRoot("src", "assets", "Raleway.ttf"),
-			"https://fonts.googleapis.com/css?family=Raleway:300,400,500,700&display=swap"
-		]
+		site: "ArkType"
+		// font: {
+		// 	title: {
+		// 		families: ["Raleway"],
+		// 		weight: "Bold",
+		// 		size: 100
+		// 	},
+		// 	description: {
+		// 		families: ["Raleway"],
+		// 		weight: "SemiBold",
+		// 		size: 40
+		// 	}
+		// }
+		// fonts: [
+		// 	fromPackageRoot("src", "assets", "Raleway.ttf"),
+		// 	"https://fonts.googleapis.com/css?family=Raleway:300,400,500,700&display=swap"
+		// ]
 	})
 )
